+++ devel/scripts/check_NMdata.R
@@ -1,0 +1,6 @@
+library(devtools)
+
+load_all()
+
+res.test <- test()
+res.chek <- check()